requirements.txt
fastapi[all]
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
pydantic-settings
email-validator
slowapi
pytest
httpx
